+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+  - Edge detail page - add container monitoring stats
+
+
 ## [2.18.0] - 2021-05-06
 
 ### Changed
@@ -16,7 +21,6 @@
     doesn't set api-credentials ttl
   - Edge page - Creation of nuvlabox-installation-trigger-usb.nuvla
     creation data content fix
-  - Edge detail page - add container monitoring stats
 
 ## [2.17.2] - 2021-04-28